This commit was generated by cvs2svn to compensate for changes in r5795, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@5796 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/drive/riot1drive0.c
@@ -70,6 +70,7 @@
 #include "iecdrive.h"
 #include "parallel.h"
 #include "riotd.h"
+#include "types.h"
 
 static void undump_pra(BYTE byte)
 {
@@ -86,7 +87,7 @@ static void undump_prb(BYTE byte)
 
 inline static void store_prb(BYTE byte)
 {
-    parallel_drv0_set_bus(parallel_atn ? 0xff : byte);
+    parallel_drv0_set_bus((BYTE)(parallel_atn ? 0xff : byte));
 }
 
 void riot1d0_set_pardata(void)
